Add Applicant Class
The applicant class is essentially the same as the previous person class.

Refactoring the Person class to the Applicant class better reflects our application's data model, and builds the foundation for adding more features, such as the Job class.

Testing for the Applicant class is also added, and the tests are still essentially the same as the Person class. To aid in the testing, the classes TypicalApplicants and ApplicantBuilder have been created, but without AddressBook dependencies. That must be added in a future commit/PR.