Fix JSON Property Person to Applicant
This change allows the read and save tests for LinkedOUT to pass.

The reason the tests were failing is because the JSON reads a person object and cannot load a person object into the applicant object. Likewise when saving to storage, the expected value is a person not an applicant. Changing the test file changes expected values to applicant.