Fix index bug for flag command
There was a bug where if the number provided in the input was too large, the wrong error message will be shown, particularly for commands that make use of indexes.

This problem was caused by bad error handling. If a ParseException is thrown by ParserUtil, flag command (and any of the other commands for that matter), will catch that exception and rethrow a generic exception.

Flag command now adds the underlying exception message from ParserUtil with the the generic error message.